+++ packages/google_maps_flutter/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapController.java
@@ -29,6 +29,7 @@ import com.google.android.gms.maps.model.CameraPosition;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.LatLngBounds;
 import com.google.android.gms.maps.model.Marker;
+import com.google.android.gms.maps.model.Polyline;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.PluginRegistry;
@@ -45,10 +46,19 @@ import com.google.maps.android.clustering.ClusterItem;
 import com.google.maps.android.clustering.ClusterManager;
 
 /** Controller of a single GoogleMaps MapView instance. */
-final class GoogleMapController implements Application.ActivityLifecycleCallbacks, GoogleMap.OnCameraIdleListener,
-    GoogleMap.OnCameraMoveListener, GoogleMap.OnCameraMoveStartedListener, GoogleMap.OnInfoWindowClickListener,
-    GoogleMap.OnMarkerClickListener, GoogleMapOptionsSink, MethodChannel.MethodCallHandler, OnMapReadyCallback,
-    GoogleMap.OnMapClickListener, PlatformView {
+final class GoogleMapController
+    implements Application.ActivityLifecycleCallbacks,
+        GoogleMap.OnCameraIdleListener,
+        GoogleMap.OnCameraMoveListener,
+        GoogleMap.OnCameraMoveStartedListener,
+        GoogleMap.OnInfoWindowClickListener,
+        GoogleMap.OnMarkerClickListener,
+        GoogleMap.OnPolylineClickListener,
+        GoogleMapOptionsSink,
+        MethodChannel.MethodCallHandler,
+        OnMapReadyCallback,
+        GoogleMap.OnMapClickListener,
+        PlatformView {
 
   private static final String TAG = "GoogleMapController";
   private final int id;
@@ -56,6 +66,7 @@ final class GoogleMapController implements Application.ActivityLifecycleCallback
   private final MethodChannel methodChannel;
   private final PluginRegistry.Registrar registrar;
   private final MapView mapView;
+  private final Map<String, PolylineController> polylines;
   private GoogleMap googleMap;
   private boolean trackCameraPosition = false;
   private boolean myLocationEnabled = false;
@@ -70,6 +81,9 @@ final class GoogleMapController implements Application.ActivityLifecycleCallback
   private List<Object> initialClusterItems;
   private ClusterManager<ClusterItemController> clusterManager;
   private MarkerManager markerManager;
+  private final PolylinesController polylinesController;
+  private List<Object> initialMarkers;
+  private List<Object> initialPolylines;
 
   GoogleMapController(int id, Context context, AtomicInteger activityState, PluginRegistry.Registrar registrar,
       GoogleMapOptions options) {
@@ -78,6 +92,7 @@ final class GoogleMapController implements Application.ActivityLifecycleCallback
     this.activityState = activityState;
     this.registrar = registrar;
     this.mapView = new MapView(context, options);
+    this.polylines = new HashMap<>();
     this.density = context.getResources().getDisplayMetrics().density;
     methodChannel = new MethodChannel(registrar.messenger(), "plugins.flutter.io/google_maps_" + id);
     methodChannel.setMethodCallHandler(this);
@@ -85,6 +100,7 @@ final class GoogleMapController implements Application.ActivityLifecycleCallback
     this.markersController = new MarkersController(methodChannel);
     this.clusterController = new ClusterController(methodChannel);
     this.markersController.setOnMarkerClickListener(this);
+    this.polylinesController = new PolylinesController(methodChannel);
   }
 
   @Override
@@ -157,9 +173,12 @@ final class GoogleMapController implements Application.ActivityLifecycleCallback
     googleMap.setOnCameraMoveListener(this);
     googleMap.setOnCameraIdleListener(this);
     googleMap.setOnMapClickListener(this);
+    googleMap.setOnPolylineClickListener(this);
+    
     updateMyLocationEnabled();
     clusterController.setGoogleMap(googleMap);
     clusterController.setClusterManager(clusterManager);
+    polylinesController.setGoogleMap(googleMap);
     googleMap.setOnCameraIdleListener(clusterManager);
     googleMap.setOnMarkerClickListener(markerManager);
     googleMap.setOnInfoWindowClickListener(clusterManager);
@@ -170,6 +189,7 @@ final class GoogleMapController implements Application.ActivityLifecycleCallback
     clusterManager.setRenderer(customRenderer);
     updateInitialMarkers();
     updateInitialClusterItems();
+    updateInitialPolylines();
   }
 
   @Override
@@ -185,7 +205,20 @@ final class GoogleMapController implements Application.ActivityLifecycleCallback
       case "map#update":
         {
           Convert.interpretGoogleMapOptions(call.argument("options"), this);
-          result.success(Convert.toJson(getCameraPosition()));
+          result.success(Convert.latlngBoundsToJson(getCameraPosition()));
+          break;
+        }
+      case "map#getVisibleRegion":
+        {
+          if (googleMap != null) {
+            LatLngBounds latLngBounds = googleMap.getProjection().getVisibleRegion().latLngBounds;
+            result.success(Convert.latlngBoundsToJson(latLngBounds));
+          } else {
+            result.error(
+                "GoogleMap uninitialized",
+                "getVisibleRegion called prior to map initialization",
+                null);
+          }
           break;
         }
       case "camera#move":
@@ -224,7 +257,17 @@ final class GoogleMapController implements Application.ActivityLifecycleCallback
           result.success(null);
           break;
         }
-
+      case "polylines#update":
+        {
+          Object polylinesToAdd = call.argument("polylinesToAdd");
+          polylinesController.addPolylines((List<Object>) polylinesToAdd);
+          Object polylinesToChange = call.argument("polylinesToChange");
+          polylinesController.changePolylines((List<Object>) polylinesToChange);
+          Object polylineIdsToRemove = call.argument("polylineIdsToRemove");
+          polylinesController.removePolylines((List<Object>) polylineIdsToRemove);
+          result.success(null);
+          break;
+        }
       case "map#isCompassEnabled":
         {
           result.success(googleMap.getUiSettings().isCompassEnabled());
@@ -266,7 +309,7 @@ final class GoogleMapController implements Application.ActivityLifecycleCallback
   @Override
   public void onMapClick(LatLng latLng) {
     final Map<String, Object> arguments = new HashMap<>(2);
-    arguments.put("position", Convert.toJson(latLng));
+    arguments.put("position", Convert.latLngToJson(latLng));
     methodChannel.invokeMethod("map#onTap", arguments);
   }
 
@@ -289,7 +332,7 @@ final class GoogleMapController implements Application.ActivityLifecycleCallback
       return;
     }
     final Map<String, Object> arguments = new HashMap<>(2);
-    arguments.put("position", Convert.toJson(googleMap.getCameraPosition()));
+    arguments.put("position", Convert.latlngBoundsToJson(googleMap.getCameraPosition()));
     methodChannel.invokeMethod("camera#onMove", arguments);
   }
 
@@ -301,6 +344,11 @@ final class GoogleMapController implements Application.ActivityLifecycleCallback
   @Override
   public boolean onMarkerClick(Marker marker) {
     return markersController.onMarkerTap(marker.getId());
+  }
+
+  @Override
+  public void onPolylineClick(Polyline polyline) {
+    polylinesController.onPolylineTap(polyline.getId());
   }
 
   @Override
@@ -456,6 +504,18 @@ final class GoogleMapController implements Application.ActivityLifecycleCallback
 
   private void updateInitialClusterItems() {
     clusterController.addClusterItems(initialClusterItems);
+  }
+          
+  @Override
+  public void setInitialPolylines(Object initialPolylines) {
+    this.initialPolylines = (List<Object>) initialPolylines;
+    if (googleMap != null) {
+      updateInitialPolylines();
+    }
+  }
+
+  private void updateInitialPolylines() {
+    polylinesController.addPolylines(initialPolylines);
   }
 
   @SuppressLint("MissingPermission")