+++ packages/google_maps_flutter/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapFactory.java
@@ -37,8 +37,13 @@ public class GoogleMapFactory extends PlatformViewFactory {
     if (params.containsKey("markersToAdd")) {
       builder.setInitialMarkers(params.get("markersToAdd"));
     }
+
     if (params.containsKey("clusterItemsToAdd")) {
       builder.setInitialClusterItems(params.get("clusterItemsToAdd"));
+    }
+    
+    if (params.containsKey("polylinesToAdd")) {
+      builder.setInitialPolylines(params.get("polylinesToAdd"));
     }
     return builder.build(id, context, mActivityState, mPluginRegistrar);
   }