+++ packages/google_maps_flutter/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapOptionsSink.java
@@ -31,4 +31,7 @@ interface GoogleMapOptionsSink {
   void setInitialMarkers(Object initialMarkers);
 
   void setInitialClusterItems(Object initialClusterItems);
+
+  void setInitialPolylines(Object initialPolylines);
+
 }