+++ packages/google_maps_flutter/example/lib/main.dart
@@ -4,18 +4,22 @@
 
 import 'package:flutter/material.dart';
 import 'animate_camera.dart';
+import 'map_coordinates.dart';
 import 'map_ui.dart';
 import 'move_camera.dart';
 import 'page.dart';
 import 'place_marker.dart';
+import 'place_polyline.dart';
 import 'scrolling_map.dart';
 
 final List<Page> _allPages = <Page>[
   MapUiPage(),
+  MapCoordinatesPage(),
   AnimateCameraPage(),
   MoveCameraPage(),
   PlaceMarkerPage(),
   ScrollingMapPage(),
+  PlacePolylinePage(),
 ];
 
 class MapsDemo extends StatelessWidget {