+++ packages/in_app_purchase/lib/src/store_kit_wrappers/enum_converters.dart
@@ -1,5 +1,6 @@
 import 'package:in_app_purchase/store_kit_wrappers.dart';
 import 'package:json_annotation/json_annotation.dart';
+import 'package:in_app_purchase/in_app_purchase_connection.dart';
 
 part 'enum_converters.g.dart';
 
@@ -17,6 +18,21 @@ class SKTransactionStatusConverter
           _$SKPaymentTransactionStateWrapperEnumMap
               .cast<SKPaymentTransactionStateWrapper, dynamic>(),
           json);
+
+  PurchaseStatus toPurchaseStatus(SKPaymentTransactionStateWrapper object) {
+    switch (object) {
+      case SKPaymentTransactionStateWrapper.purchasing:
+      case SKPaymentTransactionStateWrapper.deferred:
+        return PurchaseStatus.pending;
+      case SKPaymentTransactionStateWrapper.purchased:
+      case SKPaymentTransactionStateWrapper.restored:
+        return PurchaseStatus.purchased;
+      case SKPaymentTransactionStateWrapper.failed:
+        return PurchaseStatus.error;
+    }
+
+    throw ArgumentError('$object isn\'t mapped to PurchaseStatus');
+  }
 
   @override
   int toJson(SKPaymentTransactionStateWrapper object) =>